UI for the CV
