+++ gl-core/cartridges/CV/CV.tsx
@@ -3,137 +3,53 @@
 import * as React from 'react';
 import {
   Container,
-  Typography,
-  TextField,
-  Box,
+  AppBar,
 } from '@mui/material';
-import { Controls } from './';
-import ReactMarkdown from 'react-markdown';
-import { useSlice, useDispatch, MightyButton } from '../../';
-import { updateCVKey } from './';
+import {TailoredCV} from "./";
+import { 
+  MightyButton } from '../../';
 
 export type TCV = {
-  original?: string | null;
+  originalCV?: string | null;
 };
 
-export default function CV({ original = null }: TCV) {
-  const slice = useSlice();
-  const { cv } = slice;
-  const dispatch = useDispatch();
+export default function CV({ 
+  originalCV = null 
+}: TCV) {
 
-  const slugify = (str: string): string =>
-    str
-      .toLowerCase()
-      .trim()
-      .replace(/[^\w\s-]/g, '')
-      .replace(/\s+/g, '-');
+  // const slice = useSlice();
 
-  const extractSections = (md: string) =>
-    md
-      .split('\n')
-      .filter((line) => line.startsWith('## '))
-      .map((line) => {
-        const title = line.replace(/^## /, '').trim();
-        return {
-          id: slugify(title),
-          title,
-          visible: true,
-        };
-      });
-
-  // Initialize resume
   React.useEffect(() => {
-    const resume = cv?.resume;
-    if (!resume?.original && original) {
-      const sections = extractSections(original);
-      dispatch(
-        updateCVKey('cv.resume', {
-          original,
-          tailored: original,
-          sections,
-        })
-      );
-    }
-  }, [cv?.resume, original, dispatch]);
+  }, []);
 
-  const onJDCancel = () => {
-    dispatch(updateCVKey('cv.resume', { visible: true }));
-    dispatch(updateCVKey('cv.jd', { visible: false }));
-  }
-
-  const onAnalyse = () => {
-    console.log("onAnalyse");
-  }
-  
-
-  const resume = cv?.resume || {};
-  const jd = cv?.jd || {};
-  const visibleCV = resume.visible === true;
-  const visibleJD = jd.visible === true;
-  const displayMarkdown = resume.tailored || resume.original || original;
+  const onDownloadClick = () => {
+    console.log("onDownloadClick")
+  };
 
   return (
     <Container maxWidth="md">
-      <Controls markdown={displayMarkdown} />
-
-      {/* <pre>
-        cv: {JSON.stringify(cv, null, 2)}
-      </pre>  */}
-
-
-      {visibleCV && (
-        <Typography component="div" sx={{ mt: 4 }}>
-          <ReactMarkdown>{displayMarkdown}</ReactMarkdown>
-        </Typography>
-      )}
-
-      {visibleJD && (
-        <Box sx={{ mt: 4 }}>
-          <TextField
-            fullWidth
-            multiline
-            rows={15}
-            label="Paste Job Description here"
-            defaultValue={jd.markdown || ''}
-            onChange={(e) => {
-              dispatch(
-                updateCVKey('cv.jd', {
-                  markdown: e.target.value,
-                })
-              );
-            }}
-            onPaste={(e) => {
-              const pasted = e.clipboardData.getData('text');
-              dispatch(
-                updateCVKey('cv.jd', {
-                  markdown: pasted,
-                })
-              );
-            }}
-          />
-
-          <Box sx={{ mt: 2, display: 'flex', justifyContent: 'flex-end' }}>
-
-            <MightyButton
-              sx={{mx: 1}}
-              icon="left"
-              label="Back"
-              variant="outlined"
+      <TailoredCV markdown={originalCV}/> 
+      <AppBar 
+        position="fixed" 
+        color="primary" 
+        sx={{ 
+          background: 0,
+          boxShadow: 0,
+          top: 'auto', 
+          bottom: 0,
+          p:1,
+        }}>
+          <Container maxWidth="md">
+            <MightyButton 
+              fullWidth
+              onClick={onDownloadClick}
               color="secondary"
-              onClick={ onJDCancel }
+              label="Download PDF"
+              variant="contained"
+              icon="download"
             />
-
-            <MightyButton
-              icon="openai"
-              label="Analyse"
-              variant="outlined"
-              color="secondary"
-              onClick={ onAnalyse }
-            />
-
-          </Box>
-        </Box>
-      )}
+          </Container>
+      </AppBar>
     </Container>
   );
 }
