+++ public/globalNav.json
@@ -33,7 +33,7 @@
       "NextJS",
       "Headless CMS"
     ],
-    "excerpt": "We build and ship modern web apps for clients who need real results — fast. Custom software that works\n\n Small teams. Big output\n Real deployments from Day One\n Readable, extensible, Open Source codeb",
+    "excerpt": "We build and ship modern web apps for clients who need real results — fast. Custom software that works\n\n Small team. Big output\n Real deployments from Day One\n Readable, extensible, Open Source codeba",
     "children": [
       {
         "title": "C.V.",
@@ -46,7 +46,7 @@
           "jobs",
           "work"
         ],
-        "excerpt": "Chris Dorward\nSenior Fullstack JavaScript Developer\n\n Contact\n\nEmail: listingslab@gmail.com  \nPhone: 07745 763 122\nWeb: https://goldlabel.pro/cv  \nGitHub: https://github.com/javascriptpro  \nLinkedIn: "
+        "excerpt": "Chris Dorward\nSenior Fullstack JavaScript Developer\n\n Contact\n\nEmail:   \nWeb:   \nPhone: +44 07745 763 122  \nGitHub:   \nLinkedIn: https://www.linkedin.com/in/chrisdorward  \nWork eligibility: UK & Germa"
       },
       {
         "title": "Work",
